Affichage de la corde

+++ src/Circle.cpp
@@ -4,7 +4,7 @@ Circle::Circle() {
 }
 
 void Circle::draw(b2Body* body, sf::Color color, sf::RenderWindow* window) {
-	std::pair<int, int> dimensions = getBodyDimensions(body);
+	std::pair<float, float> dimensions = getBodyDimensions(body);
 	float radius = dimensions.first;
 	
 	b2Vec2 position = body->GetPosition();
